♻️ Refactor: 무한스크롤 재사용을 위해 커서텀 훅화(#63)
